requirements.txt
Scrapy==0.20.0
Twisted==13.2.0
cssselect==0.9.1
lxml==3.2.4
numpy==1.8.0
pyOpenSSL==0.13.1
pymongo==2.6.3
python-Levenshtein==0.10.2
queuelib==1.1.1
scikit-learn==0.14.1
scipy==0.13.1
simplejson==3.3.1
six==1.4.1
w3lib==1.5
wsgiref==0.1.2
zope.interface==4.0.5
